requirements.txt
Flask==2.0.2
Flask_Bcrypt==0.7.1
Flask_JWT_Extended==4.3.1
Flask_SQLAlchemy==2.5.1
PyMySQL==1.0.2
SQLAlchemy-Utils==0.37.9
Pillow==8.4.0
SQLAlchemy~=1.4.26
scipy~=1.7.1